fix for loop i type to size_t

+++ source/parser.cc
@@ -117,7 +117,7 @@ void split_by_pipe(vector<string> &tokens, vector<command> &cmds)
 void print_str_ascii(const std::string &input)
 {
     std::cout << "str length:" << input.size() << std::endl;
-    for (int i = 0; i < input.size(); i++)
+    for (size_t i = 0; i < input.size(); i++)
     {
         std::cout << i << " " << int(input[i]) << std::endl;
     }
